Pointer UT Fix #45. Keyboard UT Fix #46. Frame UT Fix #47. Fix #28. Tests mostly splitted. Fix #59. Moved to tests. Fix #60
