+++ src/builder.rs
@@ -22,6 +22,12 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
+/// Builder unit tests
+#[cfg(test)]
+pub(crate) mod tests {
+    include!("tests/builder.rs");
+}
+
 use crate::{frame::WindowFrame, keyboard::WindowKeyboard, pointer::WindowPointer, Window, WindowError, WindowFullScreenMode, WindowHandle, WindowManager, WindowRelativePosition, WindowSize};
 
 /// Default title of [Window].
@@ -280,187 +286,6 @@ impl WindowBuilder {
     /// Rebuild a [Window] from the [WindowBuilder] parameters.
     pub fn rebuild(&mut self, w : &mut Window) -> Result<WindowHandle, WindowError> {
         w.rebuild(self)
-    }
-
-}
-
-
-
-/*************
-* UNIT TESTS * 
-*************/
-#[cfg(test)]
-mod tests{
-    use std::io::{Read, Write};
-
-    use crate::{builder::{WB_DEFAULT_FSMODE, WB_DEFAULT_MAXIMIZED, WB_DEFAULT_MAX_SIZE, WB_DEFAULT_MINIMIZED, WB_DEFAULT_MIN_SIZE, WB_DEFAULT_PARENT, WB_DEFAULT_POSITION, WB_DEFAULT_SIZE, WB_DEFAULT_TASKBAR, WB_DEFAULT_TITLE, WB_DEFAULT_VISIBLE}, frame::{WindowFrame, WindowFrameButtonMode}, keyboard::{WindowKeyboard, WindowKeyboardMode}, pointer::{WindowCursor, WindowPointer, WindowPointerMode}, WindowBuilder, WindowFullScreenMode, WindowHandle, WindowPosition, WindowRelativePosition, WindowSize};
-
-    struct Icon {
-        index : usize,
-        content : [u8;10],
-    }
-
-    impl Read for Icon {
-        fn read(&mut self, mut buf: &mut [u8]) -> std::io::Result<usize> {
-
-            let written : usize = if buf.len() > self.content.len() - self.index {
-                self.content.len() - self.index
-            } else {
-                buf.len()
-            };
-
-            buf[0..written].copy_from_slice(&self.content[self.index..self.index+written]);
-            self.index += written;
-
-            Ok(written)
-        }
-    }
-
-    /// Test default values of WindowBuilder
-    fn test_defaults(wb : &WindowBuilder) {
-        let wkb = WindowKeyboard::new();
-        let wp = WindowPointer::new();
-        let wf = WindowFrame::new();
-
-        assert!(wb.icon == None);
-        assert!(wb.title == WB_DEFAULT_TITLE.to_string());
-        assert!(wb.min_size == WB_DEFAULT_MIN_SIZE);
-        assert!(wb.max_size == WB_DEFAULT_MAX_SIZE);
-        assert!(wb.size == WB_DEFAULT_SIZE);
-        assert!(wb.position == WB_DEFAULT_POSITION);
-        assert!(wb.parent == WB_DEFAULT_PARENT);
-        assert!(wb.fsmode == WB_DEFAULT_FSMODE);
-        assert!(wb.minimized == WB_DEFAULT_MINIMIZED);
-        assert!(wb.maximized == WB_DEFAULT_MAXIMIZED);
-        assert!(wb.visible == WB_DEFAULT_VISIBLE);
-        assert!(wb.taskbar == WB_DEFAULT_TASKBAR);
-        assert!(wb.keyboard == wkb);
-        assert!(wb.pointer == wp);
-        assert!(wb.frame == wf);
-
-    }
-
-    /// Unit tests [super::WindowBuilder] default values.
-    ///
-    /// # Verification(s)
-    /// V1 | Test each value on creation vs default values. (except pointer and keyboard that are tested on their own)
-    /// V2 | Test each pointer values.
-    /// V3 | Test each keyboard values.
-    #[test]
-    fn ut_window_builder_default() {
-        let wb = WindowBuilder::new();
-
-        // V1 | Test each value on creation vs default values. (except pointer and keyboard that are tested on their own)
-        test_defaults(&wb);
-
-    }
-
-    /// Unit tests [super::WindowBuilder] modification of values.
-    ///
-    /// # Verification(s)
-    /// V1 | Allocate value to each parameter.
-    /// V2 | Compare values VS Allocated values.
-    /// V3 | Reset and test default values.
-    #[test]
-    fn ut_window_builder_modifications() {
-        const TITLE : &str = "TEST TITLE";
-        const MIN_SIZE : WindowSize = WindowSize { width: 111, height: 222 };
-        const MAX_SIZE : WindowSize = WindowSize { width: 5555, height: 6666 };
-        const SIZE : WindowSize = WindowSize { width: 123, height: 456 };
-        const POSITION : WindowRelativePosition = WindowRelativePosition::Primary(WindowPosition { x: 987, y: 345 });
-
-        const PARENTHANDLE : usize = 1;
-        const PARENT : Option<WindowHandle> = Some(&PARENTHANDLE);
-        const FSMODE : Option<WindowFullScreenMode> = Some(WindowFullScreenMode::Primary);
-
-        const WKB_MODE : WindowKeyboardMode = WindowKeyboardMode::Text;
-        const WKB_REPEAT : bool = true;
-
-        const WP_MODE : WindowPointerMode = WindowPointerMode::Acceleration;
-        const WP_CONFINED : bool = true;
-        const WP_VISIBLE : bool = false;
-        const WP_CURSOR : WindowCursor = WindowCursor::Move;
-
-        let mut icon = Icon{ index:0, content: [0,1,2,3,4,5,6,7,8,9] };
-
-        let mut wkb = WindowKeyboard::new();
-        wkb.mode = WKB_MODE;
-        wkb.auto_repeat = WKB_REPEAT;
-
-        let mut wp = WindowPointer::new();
-        wp.mode = WP_MODE;
-        wp.confined = WP_CONFINED;
-        wp.visible = WP_VISIBLE;
-        wp.cursor = WP_CURSOR;
-
-        let mut wf = WindowFrame::new();
-        wf.visible = false;
-        wf.resizable = false;
-        wf.min_button = WindowFrameButtonMode::Disable;
-        wf.max_button = WindowFrameButtonMode::Hidden;
-        wf.close_button = WindowFrameButtonMode::Overriden;
-
-        // V1 | Allocate value to each parameter.
-        let mut wb = WindowBuilder::new();
-                wb.title(TITLE)
-                .icon(Some(&mut icon))
-                .size_min(MIN_SIZE)
-                .size_max(MAX_SIZE)
-                .size(SIZE)
-                .position(POSITION)
-                .pointer(wp)
-                .keyboard(wkb)
-                .frame(wf)
-                .parent(PARENT)
-                .fullscreen(FSMODE)
-                .minimize()
-                .maximize()
-                .hide()
-                .taskbar(false);
-
-        
-        // V2 | Compare values VS Allocated values.
-        let mut wkb = WindowKeyboard::new();
-        wkb.mode = WKB_MODE;
-        wkb.auto_repeat = WKB_REPEAT;
-
-        let mut wp = WindowPointer::new();
-        wp.mode = WP_MODE;
-        wp.confined = WP_CONFINED;
-        wp.visible = WP_VISIBLE;
-        wp.cursor = WP_CURSOR;
-
-        let mut wf = WindowFrame::new();
-        wf.visible = false;
-        wf.resizable = false;
-        wf.min_button = WindowFrameButtonMode::Disable;
-        wf.max_button = WindowFrameButtonMode::Hidden;
-        wf.close_button = WindowFrameButtonMode::Overriden;
-        
-        assert!(wb.icon.is_some());
-        if let Some(icon) = &wb.icon {
-            assert!(icon.len() == 10);
-        }
-        
-        assert!(wb.title == TITLE.to_string());
-        assert!(wb.min_size == MIN_SIZE);
-        assert!(wb.max_size == MAX_SIZE);
-        assert!(wb.size == SIZE);
-        assert!(wb.position == POSITION);
-        assert!(wb.parent == PARENT);
-        assert!(wb.fsmode == FSMODE);
-        assert!(wb.minimized == true);
-        assert!(wb.maximized == true);
-        assert!(wb.visible == false);
-        assert!(wb.keyboard == wkb);
-        assert!(wb.pointer == wp);
-        assert!(wb.frame == wf);
-        assert!(!wb.taskbar);
-
-        // V3 | Reset and test default values.
-        wb.reset();
-        test_defaults(&wb);
-
     }
 
 }
