+++ src/tests/README.md
@@ -1,0 +1,3 @@
+# Unit tests folder
+
+Contains unit tests
