+++ src/window.rs
@@ -23,10 +23,20 @@ SOFTWARE.
 */
 
 
+#[cfg(test)]
+mod tests{
+    // Window::get_window_desktop_position() unit tests
+    include!("tests/gwdp.rs");
+
+    // Window unit tests
+    include!("tests/window.rs");
+}
+
+
 
 use std::rc::Rc;
 
-use crate::{ display::{self, Desktop, Display, DisplayDesktopPosition, DisplayHandle, DisplayResolution, Displays}, frame::WindowFrame, keyboard::WindowKeyboard, linux::pointer, pointer::WindowPointer, WindowBuilder, WindowError};
+use crate::{ display::{Desktop, DisplayHandle, DisplayResolution, Displays}, frame::WindowFrame, keyboard::WindowKeyboard, pointer::WindowPointer, WindowBuilder, WindowError};
 
 /// Window handle used by the [WindowManager](crate::WindowManager).
 /// 
@@ -604,369 +614,6 @@ impl Window {
         
 
         
-    }
-
-}
-
-
-
-
-/*************
-* UNIT TESTS * 
-*************/
-
-/// display unit tests
-#[cfg(test)]
-mod tests{
-    use std::rc::Rc;
-
-    use crate::{display::{tests::create_displays, Displays}, Window, WindowError, WindowPosition, WindowRelativePosition, WindowSize};
-
-
-    fn assert_position(position : &WindowPosition, expected_x : i32, expected_y : i32) {
-        assert!(position.x == expected_x && position.y == expected_y, "Expected ({},{}), got ({},{})", expected_x, expected_y, position.x, position.y);
-    }
-
-    /// Unit tests get_window_desktop_position
-    ///
-    /// # Verification(s)
-    /// V1 | WindowRelativePosition::Desktop
-    /// V2 | WindowRelativePosition::DesktopCenter
-    /// V3 | WindowRelativePosition::Display
-    /// V4 | WindowRelativePosition::DisplayCenter
-    /// V5 | WindowRelativePosition::Parent
-    /// V6 | WindowRelativePosition::ParentCenter
-    /// V7 | WindowRelativePosition::Primary
-    /// V8 | WindowRelativePosition::PrimaryCenter
-    #[test]
-    fn ut_get_window_desktop_position() {
-        let displays : Rc<Displays> = Rc::new(create_displays(true));
-        let size = WindowSize { width: 800, height: 600 };
-        let pps : Option<(WindowPosition, WindowSize)> = Some((WindowPosition { x: 400, y: 150 }, WindowSize { width: 1024, height: 768 }));
-       
-        // V1 | WindowRelativePosition::Desktop
-        let position = WindowRelativePosition::Desktop(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, None, true) {
-            Ok(position) => assert_position(&position, 50, 125),
-            Err(_) => assert!(false),
-        }
-
-        // V2 | WindowRelativePosition::DesktopCenter
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DesktopCenter, None, true) {
-            Ok(position) => assert_position(&position, 560, 780),
-            Err(_) => assert!(false),
-        }
-
-        // V3 | WindowRelativePosition::Display
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(100, 200)), None, true) {
-            Ok(position) => assert!(position.x == 100 && position.y == 1280, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(1, WindowPosition::new(100, 200)), None, true) {
-            Ok(position) => assert!(position.x == 100 && position.y == 200, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(2, WindowPosition::new(100, 200)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::DisplayInvalidHandle)"),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // V4 | WindowRelativePosition::DisplayCenter
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DisplayCenter(0), None, true) {
-            Ok(position) => assert!(position.x == 560 && position.y == 1320, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DisplayCenter(1), None, true) {
-            Ok(position) => assert!(position.x == 560 && position.y == 240, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DisplayCenter(2), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::DisplayInvalidHandle)"),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // V5 | WindowRelativePosition::Parent
-        // Without parent fallback to desktop
-        let position = WindowRelativePosition::Parent(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, None, true) {
-            Ok(position) => assert_position(&position, 50, 125),
-            Err(_) => assert!(false),
-        }
-
-        // With parent
-        let position = WindowRelativePosition::Parent(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, pps, true) {
-            Ok(position) => assert_position(&position, 450, 275),
-            Err(_) => assert!(false),
-        }
-
-        // V6 | WindowRelativePosition::ParentCenter
-
-        // Without parent fallback to desktop
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::ParentCenter, None, true) {
-            Ok(position) => assert_position(&position, 560, 780),
-            Err(_) => assert!(false),
-        }
-        
-        // With parent
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::ParentCenter, pps, true) {
-            Ok(position) => assert_position(&position, 512, 234),
-            Err(_) => assert!(false),
-        }
-
-        // V7 | WindowRelativePosition::Primary
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Primary(WindowPosition::new(255, 311)), None, true) {
-            Ok(position) => assert!(position.x == 255 && position.y == 1391, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-
-        // V8 | WindowRelativePosition::PrimaryCenter
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::PrimaryCenter, None, true) {
-            Ok(position) => assert!(position.x == 560 && position.y == 1320, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-
-    }
-
-    /// Unit tests get_window_desktop_position OOB
-    ///
-    /// # Verification(s)
-    /// V1 | WindowError::WindowRelativePositionOOB for each type with position.
-    /// V2 | WindowError::WindowRelativePositionOOB unchecked for each type with position.
-    #[test]
-    fn ut_get_window_desktop_position_oob() {
-        let displays : Rc<Displays> = Rc::new(create_displays(true));
-        let size = WindowSize { width: 800, height: 600 };
-        let pps : Option<(WindowPosition, WindowSize)> = Some((WindowPosition { x: 400, y: 150 }, WindowSize { width: 1024, height: 768 }));
-
-        // V1 | WindowError::WindowRelativePositionOOB for each type with position.
-        // Desktop
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(i32::MIN / 2, 0)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(i32::MAX / 2, 0)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(0, i32::MIN / 2)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(0, i32::MAX / 2)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // Display
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(i32::MIN / 2, 0)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(1, WindowPosition::new(i32::MAX / 2, 0)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(0, i32::MIN / 2)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(1, WindowPosition::new(0, i32::MAX / 2)), None, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // Parent
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(i32::MIN / 2, 0)), pps, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(i32::MAX / 2, 0)), pps, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(0, i32::MIN / 2)), pps, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(0, i32::MAX / 2)), pps, true) {
-            Ok(_) => assert!(false, "Expected Err(WindowError::WindowRelativePositionOOB)"),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-    
-        // V2 | WindowError::WindowRelativePositionOOB unchecked for each type with position.
-        // Desktop
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(i32::MIN / 2, 0)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(i32::MAX / 2, 0)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(0, i32::MIN / 2)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Desktop(WindowPosition::new(0, i32::MAX / 2)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-
-        // Display
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(i32::MIN / 2, 0)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(1, WindowPosition::new(i32::MAX / 2, 0)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(0, i32::MIN / 2)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(1, WindowPosition::new(0, i32::MAX / 2)), None, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-
-        // Parent
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(i32::MIN / 2, 0)), pps, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(i32::MAX / 2, 0)), pps, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(0, i32::MIN / 2)), pps, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Parent(WindowPosition::new(0, i32::MAX / 2)), pps, false) {
-            Ok(_) => assert!(true),
-            Err(err) => assert!(false, "Should not be OOB"),
-        }
-
-    }
-
-    /// Unit tests get_window_desktop_position limit
-    ///
-    /// # Verification(s)
-    /// V1 | Test each type with no displays or desktop
-    #[test]
-    fn ut_get_window_desktop_position_limit() {
-        let displays : Rc<Displays> = Rc::new(Displays { list: Vec::new(), desktop: None });
-        let size = WindowSize { width: 800, height: 600 };
-        let pps : Option<(WindowPosition, WindowSize)> = Some((WindowPosition { x: 400, y: 150 }, WindowSize { width: 1024, height: 768 }));
-
-        // WindowRelativePosition::Desktop
-        let position = WindowRelativePosition::Desktop(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-        // unchecked
-        let position = WindowRelativePosition::Desktop(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, None, false) {
-            Ok(position) => assert_position(&position, 50, 125),
-            Err(_) => assert!(false),
-        }
-
-        // WindowRelativePosition::DesktopCenter
-        // checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DesktopCenter, None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // unchecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DesktopCenter, None, false) {
-            Ok(position) => assert_position(&position, 0, 0),
-            Err(_) => assert!(false),
-        }
-
-        // WindowRelativePosition::Display
-        // Checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(100, 200)), None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // UnChecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Display(0, WindowPosition::new(100, 200)), None, false) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // WindowRelativePosition::DisplayCenter
-        // Checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DisplayCenter(0), None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // UnChecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::DisplayCenter(0), None, false) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::DisplayInvalidHandle),
-        }
-
-        // WindowRelativePosition::Parent
-
-        // Checked
-        let position = WindowRelativePosition::Parent(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, pps, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // UnChecked
-        let position = WindowRelativePosition::Parent(WindowPosition { x: 50, y: 125 });
-        match Window::get_window_desktop_position(size, displays.clone(), position, pps, false) {
-            Ok(position) => assert_position(&position, 450, 275),
-            Err(_) => assert!(false),
-        }
-
-        // WindowRelativePosition::ParentCenter
-        // Checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::ParentCenter, pps, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // UnChecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::ParentCenter, pps, false) {
-            Ok(position) => assert_position(&position, 512, 234),
-            Err(_) => assert!(false),
-        }
-
-        // WindowRelativePosition::Primary
-        // Checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Primary(WindowPosition::new(255, 311)), None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // UnChecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::Primary(WindowPosition::new(255, 311)), None, false) {
-            Ok(position) => assert!(position.x == 255 && position.y == 311, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
-
-        // WindowRelativePosition::PrimaryCenter
-        // Checked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::PrimaryCenter, None, true) {
-            Ok(_) => assert!(false),
-            Err(err) => assert!(err == WindowError::WindowRelativePositionOOB),
-        }
-
-        // Unchecked
-        match Window::get_window_desktop_position(size, displays.clone(), WindowRelativePosition::PrimaryCenter, None, false) {
-            Ok(position) => assert!(position.x == 0 && position.y == 0, " Result={:?}", position),
-            Err(_) => assert!(false),
-        }
     }
 
 }
